Update README.md
First, there is no "free" service option for the NLC, so that should be changed to "standard".

Second, The tutorial says to change the classifier ID in the app.js file. However, the file already checks for the cf environment variable CLASSIFIER_ID. So I recommend instructing users to do `cf set-env` instead of modifying the source code.

If we use the above method to set env, we need to push the application before setting the environment variable, since the environment variables are bound to the app.